Commit project 11 - camera v12 - verFoto

+++ app/src/main/java/com/example/ejt7/models/Cine.kt
@@ -3,6 +3,6 @@ package com.example.ejt7.models
 import android.os.Parcel
 import android.os.Parcelable
 
-class Cine (var id:Long, var nombreCine:String, var ciudad:Ciudad, var latitud:Double, var longitud:Double ) {
+class Cine (var id:Int, var nombreCine:String, var ciudad:Ciudad, var latitud:Double, var longitud:Double ) {
 
 }
