+++ app/src/main/java/com/example/ejt7/models/Pelicula.kt
@@ -1,5 +1,5 @@
 package com.example.ejt7.models
 
 
-class Pelicula (var id:Long,var title:String,var description:String, var poster:Int, var time:Int, var year:Int, var country:String, var uri:String){
+class Pelicula (var id:Int,var title:String,var description:String, var poster:Int, var time:Int, var year:Int, var country:String, var uri:String){
 }
